Migrate electron-app.js to electron-app.ts and updater.js to updater.ts. Also create tasks to compile code with Gulp.
