+++ package.json
@@ -18,6 +18,7 @@
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^6.0.2",
     "css-loader": "^3.6.0",
+    "del": "^5.1.0",
     "electron": "^9.0.4",
     "electron-builder": "^22.7.0",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
@@ -28,6 +29,7 @@
     "gulp-imagemin": "^7.1.0",
     "gulp-pretty-url": "^0.1.1",
     "gulp-sass": "^4.1.0",
+    "gulp-typescript": "^6.0.0-alpha.1",
     "har-validator": "^5.1.3",
     "html-webpack-plugin": "^4.3.0",
     "jest": "^26.0.1",
